refactor: serialization separated from frame module
