+++ src/socket_server/mod.rs
@@ -3,3 +3,4 @@ pub mod frame;
 pub mod server;
 pub mod receiver;
 pub mod queue_manager;
+pub mod deserialization;
