+++ src/socket_server/server.rs
@@ -1,4 +1,5 @@
-use crate::socket_server::frame::{Frame, deserialize_frame};
+use crate::socket_server::frame::Frame;
+use crate::socket_server::deserialization::deserialize_frame;
 use crossbeam_channel::Sender;
 use std::io::BufReader;
 use std::os::unix::net::UnixListener;
